requirements.txt
# Requirements for the project

# 核心資料處理與金融資料
yfinance
finlab
pandas
numpy

# 設定與環境變數
pyyaml
python-dotenv

# 測試
pytest

# Google Cloud
google-cloud-storage
google-cloud-bigquery

# Parquet / BigQuery 相容
pyarrow

# 因子分析
alphalens-reloaded
matplotlib

# 機器學習
scikit-learn